update UI for read only superadmins

+++ back/server/superadminapi/delete_group.ts
@@ -3,10 +3,14 @@ import { logError } from '../helpers/logger';
 
 export const delete_group = async (req: any, res: any): Promise<void> => {
   try {
+    if (req.session.isReadOnlySuperadmin) {
+      res.status(401).json({ error: 'Not allowed for read only superadmin' });
+      return;
+    }
     await db.query(`DELETE FROM groups WHERE id=$1`, [req.params.id]);
     res.status(200).end();
   } catch (e) {
-    logError("delete_group", e);
+    logError('delete_group', e);
     res.status(400).end();
   }
 };
